Added functions for password, server , upload and delete
Added 4 functions
get_password - This will prompt the user for password and the same will
be used for all requested operations in that single run
get_server - This will parse the output from the command to get the
server, if there is an error program will exit
upload - This will read the temp upload list file built from config file
and do the upload. Will not proceed if there are no upload file/folders.
delete - This will read the temp delete list file built from config file
and do the delete. Will not proceed if there are no delete file/folders.