Added variable for keys in config and fixed bug for no upload/delete list
Added indivdual variable for the key names used in the config list,
easier to make any edits in the future.

Also fixed a bug where if the upload or delete list is empty or missing
in the config file the program will create files with a single empty
line. This was causing unwanted behaviour during upload.

Added a '--' to indicate prmopt while requesting for password