Added ACL_BACKUP logic and minor refactor for config file param parsing
Added a logic wherin user can provide a path in teh config file against
ACL_BACKUP where the backed up files and folder's ACL info will be
written using getfacl. This will create a unique file every run for the
items in that run. This is needed as on restore the
permissions/owner/group info is not restored. IDrive also suggest to
backup this metadat info and use the same while restoring. setfacl
command can be used to restore after amending the path in the file.
http://www.idrive.com/faq_linux.htm

Minor refactoring of the commands used to parse the config params by
adding a -s option to the cut command which will prevent the key name
being assigned as value when there is no value maintained and removed
the code used to handle this case.