Added download function and changed G option
Added the download function. -g with argument is mandatory. The comma
separated file/folder paths wrt to IDrive will be provided here. This
will then be converted to a temp file. With -G a download location can
be specified, if not given then a location in /tmp will be used.