Added help file, template .idrivewrc, updated show_help, added check for external programs before usage and split out init from read_config_init
Added a help file and the show_help function has been updated to read
this file for both minimal help and detailed help

Added a .idrivewrc file as a template, this will be made available in
the /usr/share/doc/idrive-wrapper/ folder to help users

Fixed a bug where if the value for the IDRIVE_HOME_FOLDER in config
does not have a prefix '/' then upload/delete fails.

Separated the init for non config variables to a init_script function
and renamed read_config_init to read_config. read_config will only
deal with the confi file.

Included a check for getfacl and more programs in init_script and they
will be used only if available in system.