Added log file functionality(o option),changed verison to 0.3b and updated manual
Added functionality to log to a file using the -o option. Writing to
terminal will still happen irrespective of logging option. Using the tee
command for this. If no write access is available or file path is
invalid then no logs will be written. output of idevsutil both 1 and 2
are also directed to the files using tee.

Changed version to 0.3 beta

Updated the manual