Look for helpfile in current dir, updated version to 1.0, updated manual and Readme.md. Will be first stable version.
The helpfile will now be looked for in the current directory as well.
This is to allow other distros to use the app without root privileges.

Version has been updated to 1.0 in source and manual

readme.md has been updated with instructions for installation Arch and other distros
and usage instructions.